protect routes with auth middleware

+++ backend/src/controllers/notes.ts
@@ -2,10 +2,16 @@ import { RequestHandler } from 'express';
 import NoteModal from '../models/note';
 import createHttpError from 'http-errors';
 import mongoose from 'mongoose';
+import { assertIsDefined } from '../util/assertIsDefined';
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+    const authenticatedUser = req.session.userId;
+
     try {
-        const notes = await NoteModal.find().exec();
+        assertIsDefined(authenticatedUser);
+        const notes = await NoteModal.find({
+            userId: authenticatedUser,
+        }).exec();
         res.status(200).json(notes);
     } catch (error) {
         next(error);
@@ -25,8 +31,10 @@ export const createNote: RequestHandler<
 > = async (req, res, next) => {
     const title = req.body.title;
     const text = req.body.text;
+    const authenticatedUser = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUser);
         if (!title) {
             throw createHttpError(400, 'Title is required');
         }
@@ -34,6 +42,7 @@ export const createNote: RequestHandler<
         const newNote = await NoteModal.create({
             title: title,
             text: text,
+            userId: authenticatedUser,
         });
 
         res.status(201).json(newNote);
@@ -44,7 +53,9 @@ export const createNote: RequestHandler<
 
 export const getNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticatedUser = req.session.userId;
     try {
+        assertIsDefined(authenticatedUser);
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, 'Invalid note id');
         }
@@ -52,6 +63,10 @@ export const getNote: RequestHandler = async (req, res, next) => {
 
         if (!note) {
             throw createHttpError(404, 'Note not found');
+        }
+
+        if (!note.userId.equals(authenticatedUser)) {
+            throw createHttpError(403, 'Not allowed');
         }
 
         res.status(200).json(note);
@@ -78,7 +93,10 @@ export const updateNote: RequestHandler<
     const noteId = req.params.noteId;
     const newTitle = req.body.title;
     const newText = req.body.text;
+    const authenticatedUser = req.session.userId;
     try {
+        assertIsDefined(authenticatedUser);
+
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, 'Invalid note id');
         }
@@ -91,6 +109,10 @@ export const updateNote: RequestHandler<
 
         if (!note) {
             throw createHttpError(404, 'Note not found');
+        }
+
+        if (!note.userId.equals(authenticatedUser)) {
+            throw createHttpError(403, 'Not allowed');
         }
 
         note.title = newTitle;
@@ -106,8 +128,11 @@ export const updateNote: RequestHandler<
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticatedUser = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUser);
+
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, 'Invalid note id');
         }
@@ -116,6 +141,10 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
 
         if (!note) {
             throw createHttpError(404, 'Note not found');
+        }
+
+        if (!note.userId.equals(authenticatedUser)) {
+            throw createHttpError(403, 'Not allowed');
         }
 
         await note.deleteOne();
